2021 Computer Application Ability Competition/requirements.txt
absl-py==0.10.0
affinegap==1.11
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.4
argh==0.26.2
asn1crypto==1.3.0
astropy==4.0.1.post1
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
bkcharts==0.2
bleach==3.1.5
boto==2.49.0
boto3==1.17.18
botocore==1.20.18
Bottleneck==1.3.2
brotlipy==0.7.0
bs4==0.0.1
BTrees==4.7.2
cachetools==4.1.1
categorical-distance==1.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
cycler==0.10.0
cytoolz==0.11.0
datetime-distance==0.1.3
decorator==4.4.2
defusedxml==0.6.0
display==1.0.0
distlib==0.3.1
docopt==0.6.2
docutils==0.16
DoubleMetaphone==0.1
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.1.0
fastcluster==1.1.28
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
glob2==0.7
gmpy2==2.0.8
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
greenlet==0.4.16
grpcio==1.32.0
h5py==2.10.0
haversine==2.3.0
HeapDict==1.0.1
highered==0.2.1
ipyparallel==6.3.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
jqdatasdk==1.8.1
json-lines==0.5.0
json5==0.9.5
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-kite==1.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.2.6
jupyterthemes==0.20.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lesscpy==0.14.0
Levenshtein-search==1.4.5
libarchive-c==2.9
locket==0.2.0
lsp==0.1.1
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.4.0
mpl-finance==0.10.1
mpmath==1.1.0
msgpack==1.0.0
multidict==4.7.6
multipledispatch==0.6.0
my-utils==0.0.1
nbconvert==5.6.1
node==0.9.25
nose==1.3.7
npm==0.1.1
numba==0.50.1
numexpr==2.7.1
numpy==1.19.5
oauthlib==3.1.0
odict==1.7.0
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.3.0
optional-django==0.1.0
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pefile==2019.4.18
pep8==1.7.1
persistent==4.6.4
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.10
pkginfo==1.5.0.1
plotly==4.12.0
pluggy==0.13.1
plumber==1.6
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycurl==7.43.0.5
pydantic==1.8
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
pyhacrf-datamade==0.2.5
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyLBFGS==0.2.0.13
PyMySQL==0.10.1
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
pyspark==3.1.1
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
QtPy==1.9.0
requests-oauthlib==1.3.0
retrying==1.3.3
rlr==2.4.5
rope==0.17.0
rsa==4.6
ruamel-yaml==0.15.87
s3transfer==0.3.4
scikit-image==0.16.2
seaborn==0.10.1
Send2Trash==1.5.0
simplecosine==1.2
simplegeneric==0.8.1
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
statsmodels==0.11.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.7.1
tornado==6.0.4
ujson==1.35
unicodecsv==0.14.1
virtualenv==20.4.2
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xgboost==1.3.3
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.5
xlwt==1.3.0
xmltodict==0.12.0
yarg==0.1.9
yarl==1.6.0
zict==2.0.0
zope.component==4.6.2
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.index==5.0.0
zope.interface==4.7.1
zope.lifecycleevent==4.3
zope.proxy==4.3.5